added stable baselines submodule
There are many parts of the current codebase that could more easily be implemented by directly changing the stable baselines sourcecode. As the goal is not just functionality but also readability, we will from now on directly change the source code. This also makes it easier to push the changes to the stable baselines 3 repository if desired